Completed sprint 3

+++ backend/User/tests/test_api.py
@@ -1,16 +1,16 @@
 from django.test import TestCase
 from django.urls import reverse
+from rest_framework.authtoken.models import Token
 from rest_framework.test import APITestCase
 from rest_framework import status
-from .models import CustomUser, Education_Field, Complete_Portfolio, PortfolioItem, AccessType
-from .serializers import CustomUserSerializer, EducationSerializer
-
+from ..models import CustomUser, Education_Field, Complete_Portfolio, PortfolioItem, AccessType
+from ..serializers import CustomUserSerializer, EducationSerializer
+from rest_framework.test import force_authenticate
 
 class CustomUserTestCase(APITestCase):
     """
     Test case for the CustomUser model and API endpoints.
     """
-
     def setUp(self):
         self.user = CustomUser.objects.create(
             username="testuser",
@@ -20,29 +20,27 @@ class CustomUserTestCase(APITestCase):
             portfolioVisibility=True,
             profileImage="http://example.com/image.jpg",
         )
-        self.education = Education_Field.objects.create(
-            user=self.user,
-            field_of_study="Computer Science",
-            major="Software Engineering",
-            minor="Data Science"
-        )
-        self.user.education = self.education
-
+        self.user.education.field_of_study = "Computer Science"
+        self.user.education.major = "Software Engineering"
+        self.user.education.minor = "Data Science"
+        self.user.education.save()
+        self.token = Token.objects.create(user=self.user)
     '''
     US 1.03 As an admin, I want to include additional information in my profile like innovation center role, so it can be visible on my profile.
     US 1.05 As a user, I want to look up other users by using the search bar, so that I can find information about my friends, organizations, and admins.
     '''
+
     def test_user_get(self):
         '''
         Test the API endpoint for retrieving a list of all users.
         '''
-        response = self.client.get(reverse('user_list'))
+        response = self.client.get(reverse('user_list'), format='json', HTTP_AUTHORIZATION=self.token.key)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data[0]['first_name'], 'Test')
         self.assertEqual(response.data[0]['last_name'], 'User')
         self.assertEqual(response.data[0]['email'], 'testuser@example.com')
         self.assertEqual(response.data[0]['portfolioVisibility'], True)
-        self.assertEqual(response.data[0]['profileImage'], None)
+        self.assertEqual(response.data[0]['profileImage'], "http://example.com/image.jpg")
         self.assertEqual(response.data[0]['education']['field_of_study'], 'Computer Science')
         self.assertEqual(response.data[0]['education']['major'], 'Software Engineering')
         self.assertEqual(response.data[0]['education']['minor'], 'Data Science')
@@ -53,11 +51,12 @@ class CustomUserTestCase(APITestCase):
     US 1.02 As a user, I want to create and manage my profile, including my basic information and academic details.   
     US 1.06 As a user, I want to be able to change the visibility of my portfolio, so that I can control if everyone/organizations/only me can see it.
     '''
+
     def test_user_patch(self):
         '''
         Test the API endpoint for updating a user's profile.
         '''
-        self.client.force_authenticate(user=self.user)
+   #     self.client.force_authenticate(user=self.user)
         response = self.client.patch(
             reverse('user_detail', kwargs={'pk': self.user.pk}),
             {
@@ -72,15 +71,18 @@ class CustomUserTestCase(APITestCase):
                     'minor': 'Updated Minor'
                 }
             },
-            format='json'
+            format='json',
+           # headers={'AUTHORIZATION':self.token.key}
+           HTTP_AUTHORIZATION = "Token " + self.token.key
+           
             # print(response.data)
         )
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data['first_name'], 'Updated')
         self.assertEqual(response.data['last_name'], 'User')
-        self.assertEqual(response.data['email'], 'testuser@example.com') # Email should not be updated
+        self.assertEqual(response.data['email'], 'testuser@example.com')  # Email should not be updated
         self.assertEqual(response.data['portfolioVisibility'], False)
-        self.assertEqual(response.data['profileImage'], None)
+        self.assertEqual(response.data['profileImage'], 'http://example.com/updated_image.jpg')
         self.assertEqual(response.data['education']['field_of_study'], 'Updated Field')
         self.assertEqual(response.data['education']['major'], 'Updated Major')
         self.assertEqual(response.data['education']['minor'], 'Updated Minor')
@@ -90,10 +92,12 @@ class CustomUserTestCase(APITestCase):
         Test the API endpoint for updating a user's profile with invalid data.
         '''
         # Invalid portfolioVisibility
+        self.client.force_authenticate(user=self.user)
         response = self.client.patch(
             reverse('user_detail', kwargs={'pk': self.user.pk}),
             {'portfolioVisibility': "OK"},  # Invalid data: portfolioVisibility should be a boolean
-            format='json'
+            format='json',
+            HTTP_AUTHORIZATION = "Token " + self.token.key
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
 
@@ -101,7 +105,8 @@ class CustomUserTestCase(APITestCase):
         response = self.client.patch(
             reverse('user_detail', kwargs={'pk': self.user.pk}),
             {'profileImage': 'invalidurl'},  # Invalid data: profileImage should be a valid URL
-            format='json'
+            format='json',
+            HTTP_AUTHORIZATION = "Token " + self.token.key
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
 
@@ -109,26 +114,26 @@ class CustomUserTestCase(APITestCase):
         response = self.client.patch(
             reverse('user_detail', kwargs={'pk': 999999}),  # Assuming 9999 is an ID that does not exist
             {'profileImage': 'invalidurl'},  # Invalid data: profileImage should be a valid URL
-            format='json'
+            format='json',
+            HTTP_AUTHORIZATION = "Token " + self.token.key
         )
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
 
     def test_user_delete(self):
         '''
         Test the API endpoint for deleting a user.
         '''
-        self.client.force_authenticate(user=self.user)
-        response = self.client.delete(reverse('user_detail', kwargs={'pk': self.user.pk}))
+        response = self.client.delete(reverse('user_detail', kwargs={'pk': self.user.pk},),HTTP_AUTHORIZATION = "Token " + self.token.key)
         self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
         self.assertEqual(CustomUser.objects.count(), 0)
         self.assertEqual(Education_Field.objects.count(), 0)
 
-    
 
 class CompletePortfolioTestCase(APITestCase):
     """
     Test case for the CompletePortfolio model and API endpoints.
     """
+
     def setUp(self):
         self.user = CustomUser.objects.create(
             username="testuser",
@@ -138,35 +143,30 @@ class CompletePortfolioTestCase(APITestCase):
             portfolioVisibility=True,
             profileImage="http://example.com/image.jpg",
         )
-        self.education = Education_Field.objects.create(
-            user=self.user,
-            field_of_study="Computer Science",
-            major="Software Engineering",
-            minor="Data Science"
-        )
-        self.user.education = self.education
-        
-    
+        self.user.education.field_of_study = "Computer Science"
+        self.user.education.major = "Software Engineering"
+        self.user.education.minor = "Data Science"
+        self.user.education.save()
+        self.token = Token.objects.create(user=self.user)
 
     def test_complete_portfolio_get(self):
         '''
         Test the API endpoint for retrieving a user's complete portfolio.
         '''
-        response = self.client.get(reverse('complete-portfolio-detail', kwargs={'user_id': self.user.pk}))
+        response = self.client.get(reverse('complete-portfolio-detail', kwargs={'user_id': self.user.pk}), HTTP_AUTHORIZATION = "Token " + self.token.key)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data['user'], self.user.pk)
         self.assertEqual(response.data['description'], "")
         self.assertEqual(response.data['items'], [])
-        
+
     def test_complete_portfolio_get_failure(self):
         '''
         Test the API endpoint for retrieving a user's complete portfolio with invalid data.
         '''
         # Trying to get a user that does not exist
-        response = self.client.get(reverse('complete-portfolio-detail', kwargs={'user_id': 999999}))
+        response = self.client.get(reverse('complete-portfolio-detail', kwargs={'user_id': 999999}), HTTP_AUTHORIZATION=self.token.key)
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
-        
     def test_complete_portfolio_patch(self):
         '''
         Test the API endpoint for updating a user's complete portfolio.
@@ -177,7 +177,8 @@ class CompletePortfolioTestCase(APITestCase):
             {
                 'description': 'Updated Description',
             },
-            format='json'
+            format='json',
+            HTTP_AUTHORIZATION = "Token " + self.token.key
         )
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data['user'], self.user.pk)
@@ -188,14 +189,17 @@ class CompletePortfolioTestCase(APITestCase):
         Test the API endpoint for updating a user's complete portfolio with invalid data.
         '''
         # Trying to add a user that does not exist
-        response = self.client.patch(reverse('complete-portfolio-detail', kwargs={'user_id': 999999}),  # Assuming 9999 is an ID that does not exist
-             {
-                 'description': 'Updated Description',
-             },
-             format='json'
-             )
+        response = self.client.patch(reverse('complete-portfolio-detail', kwargs={'user_id': 999999}),
+                                     # Assuming 9999 is an ID that does not exist
+                                     {
+                                         'description': 'Updated Description',
+                                     },
+                                     format='json',
+                                    HTTP_AUTHORIZATION = "Token " + self.token.key
+                                     )
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-        
+
+# # ---------------------------------------------------------------
 class PortfolioItemTestCase(APITestCase):
     def setUp(self):
         self.user = CustomUser.objects.create(
@@ -206,17 +210,16 @@ class PortfolioItemTestCase(APITestCase):
             portfolioVisibility=True,
             profileImage="http://example.com/image.jpg",
         )
-        self.education = Education_Field.objects.create(
-            user=self.user,
-            field_of_study="Computer Science",
-            major="Software Engineering",
-            minor="Data Science"
-        )
-        self.user.education = self.education
+        self.user.education.field_of_study = "Computer Science"
+        self.user.education.major = "Software Engineering"
+        self.user.education.minor = "Data Science"
+        self.user.education.save()
+        self.token = Token.objects.create(user=self.user)
 
-    '''
-    US 1.04 As a user, I want to upload text and link to a portfolio in my profile, so that prospective employers will see my capabilities.
-    '''
+#     '''
+#     US 1.04 As a user, I want to upload text and link to a portfolio in my profile, so that prospective employers will see my capabilities.
+#     '''
+
     def test_portfolio_item_post(self):
         '''
         Test the API endpoint for creating a new portfolio item.
@@ -230,7 +233,8 @@ class PortfolioItemTestCase(APITestCase):
                 'description': 'Description',
                 'link': 'http://example.com/link'
             },
-            format='json'
+            format='json',
+            HTTP_AUTHORIZATION = "Token " + self.token.key
         )
         # print(response.data)
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -238,9 +242,7 @@ class PortfolioItemTestCase(APITestCase):
         self.assertEqual(response.data['title'], 'Title')
         self.assertEqual(response.data['description'], 'Description')
         self.assertEqual(response.data['link'], 'http://example.com/link')
-        
-        
-        
+
     def test_complete_portfolio_get(self):
         '''
         Test the API endpoint for retrieving a user's complete portfolio.
@@ -276,15 +278,16 @@ class PortfolioItemTestCase(APITestCase):
                 'description': 'Updated Description',
                 'link': 'http://example.com/updated_link'
             },
-            format='json'
-            
+            format='json',
+            HTTP_AUTHORIZATION = "Token " + self.token.key
+
         )
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data['icon'], "http://example.com/updated_icon")
         self.assertEqual(response.data['title'], 'Updated Title')
         self.assertEqual(response.data['description'], 'Updated Description')
         self.assertEqual(response.data['link'], 'http://example.com/updated_link')
-       
+
     def test_portfolio_item_patch_failure(self):
         '''
         Test the API endpoint for updating a portfolio item with invalid data.
@@ -298,7 +301,8 @@ class PortfolioItemTestCase(APITestCase):
                 'description': 'Updated Description',
                 'link': 'http://example.com/updated_link'
             },
-            format='json'
+            format='json',
+            HTTP_AUTHORIZATION = "Token " + self.token.key
         )
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
@@ -317,21 +321,21 @@ class PortfolioItemTestCase(APITestCase):
             format='json'
         )
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-       
+
     def test_portfolio_item_delete(self):
         '''
         Test the API endpoint for deleting a portfolio item.
         '''
         self.test_portfolio_item_post()
         self.assertEqual(PortfolioItem.objects.count(), 1)
-        response = self.client.delete(reverse('portfolio-item-detail', kwargs={'pk': 1}))
+        response = self.client.delete(reverse('portfolio-item-detail', kwargs={'pk': 1}), HTTP_AUTHORIZATION = "Token " + self.token.key)
         self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
         self.assertEqual(PortfolioItem.objects.count(), 0)
         response = self.client.get(reverse('complete-portfolio-detail', kwargs={'user_id': self.user.pk}))
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data['items'], [])
-   
-    
+
+
 class AccessTypeTestCase(APITestCase):
     def test_access_type_get(self):
         '''
@@ -349,13 +353,13 @@ class AccessTypeTestCase(APITestCase):
         self.assertEqual(response.data[1]['name'], 'Another Access Type')
 
 
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
+
+
+
+
+
+
+
+
+
+
